requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.0.8
qdrant-client==1.11.3
sentence-transformers==3.0.1
numpy==1.26.4

# тесты
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
